requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Anthropic Claude SDK
anthropic==0.43.0

# HTTP Client
requests==2.32.3

# Environment Variables
python-dotenv==1.0.1

# File Upload Support
python-multipart==0.0.20

# MCP SDK
mcp==1.3.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-mock==3.14.0
responses==0.25.3

# Type Hints
pydantic==2.10.5
